Format corrections and move tree checker into utilities.
